Remove useless header
